refactor(app): Restructure flat app package into domain, services, and integrations zones
Pure move, no behavior change: split app/domain.py into
domain/provenance.py and domain/readiness.py; move ledger and
sanitization into domain/ (pure, no I/O) so dependency edges stay
one-way; split the 551-line app/fortyguard.py along its fault lines
into integrations/fortyguard/{errors,transport,client,contracts}.py;
move cache/execution into services/ and the usage utility beside the
provider stack. Update all imports in tests and scripts; setuptools
now discovers app* sub-packages. Preparatory restructure for issue
#10 per the extraction manifest's suggested hierarchy.
