+++ app/domain/trip.py
