+++ app/integrations/fortyguard/contracts.py
@@ -1,23 +1,17 @@
-"""Offline-safe contracts for the asynchronous FortyGuard API."""
+"""Offline-safe request and response contracts for the FortyGuard API."""
 
 from __future__ import annotations
 
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from datetime import date, datetime, timedelta
 from enum import Enum
 import math
-import re
-import json
-from time import sleep as default_sleep
-from typing import Callable, Mapping, Protocol, Sequence
-from urllib.error import HTTPError, URLError
-import socket
-from urllib.request import Request, urlopen
+from typing import Mapping, Sequence
+
 from shapely.geometry import shape
 
-from app.domain import Provenance
-from app.ledger import CreditLedger, UsageRecord
-from app.security import sanitize_payload
+from app.domain.provenance import Provenance
+from app.domain.security import sanitize_payload
 
 
 class AnalyticType(str, Enum):
@@ -156,256 +150,6 @@ def normalize_env_params_response(
         False,
         "caller-supplied temperature anchor; not a real 24-hour forecast",
     )
-
-
-class ProviderErrorKind(str, Enum):
-    AUTHENTICATION = "authentication"
-    VALIDATION = "validation_or_plan"
-    RATE_LIMIT = "rate_limit"
-    SERVER = "server"
-    MALFORMED_RESPONSE = "malformed_response"
-    TASK_FAILURE = "task_failure"
-    TIMEOUT = "timeout"
-
-
-@dataclass(frozen=True)
-class ProviderError(Exception):
-    kind: ProviderErrorKind
-    status_code: int | None = None
-    detail: str = "provider request failed"
-
-    def __str__(self) -> str:
-        return f"{self.kind.value}: {self.detail}"
-
-
-def _sanitize_detail(detail: str) -> str:
-    detail = re.sub(r"(?i)(api[_ -]?key|authorization|token)\s*[:=]\s*[^\s,;]+", r"\1=[redacted]", detail)
-    return detail[:160]
-
-
-def classify_provider_error(status_code: int | None, detail: str = "") -> ProviderError:
-    if status_code == 408:
-        kind = ProviderErrorKind.TIMEOUT
-    elif status_code in (401, 403):
-        kind = ProviderErrorKind.AUTHENTICATION
-    elif status_code in (400, 404, 422):
-        kind = ProviderErrorKind.VALIDATION
-    elif status_code == 429:
-        kind = ProviderErrorKind.RATE_LIMIT
-    elif status_code is not None and status_code >= 500:
-        kind = ProviderErrorKind.SERVER
-    else:
-        kind = ProviderErrorKind.MALFORMED_RESPONSE
-    return ProviderError(kind, status_code, _sanitize_detail(detail) or "provider request failed")
-
-
-class FortyGuardTransport(Protocol):
-    def post(self, endpoint: str, payload: Mapping[str, object], api_key: str) -> Mapping[str, object]: ...
-
-    def get(self, endpoint: str, api_key: str) -> Mapping[str, object]: ...
-
-
-class HttpFortyGuardTransport:
-    class HttpError(Exception):
-        def __init__(self, response: object) -> None:
-            self.response = response
-
-    def __init__(
-        self,
-        base_url: str,
-        *,
-        timeout_seconds: float = 15,
-        opener: Callable[..., object] = urlopen,
-    ) -> None:
-        self.base_url = base_url.rstrip("/")
-        self.timeout_seconds = timeout_seconds
-        self._opener = opener
-
-    def post(self, endpoint: str, payload: Mapping[str, object], api_key: str) -> Mapping[str, object]:
-        return self._request(endpoint, api_key, payload)
-
-    def get(self, endpoint: str, api_key: str) -> Mapping[str, object]:
-        return self._request(endpoint, api_key)
-
-    def _request(
-        self, endpoint: str, api_key: str, payload: Mapping[str, object] | None = None
-    ) -> Mapping[str, object]:
-        request = Request(
-            f"{self.base_url}/{endpoint.lstrip('/')}",
-            data=json.dumps(payload).encode() if payload is not None else None,
-            headers={"X-API-Key": api_key, "Content-Type": "application/json"},
-            method="POST" if payload is not None else "GET",
-        )
-        try:
-            with self._opener(request, self.timeout_seconds) as response:  # type: ignore[attr-defined]
-                parsed = json.loads(response.read())
-        except self.HttpError as error:
-            status = getattr(error.response, "status", None)
-            if payload is None and (status in (404, 408, 429) or isinstance(status, int) and status >= 500):
-                return {"status_code": status}
-            raise classify_provider_error(status, "provider HTTP request failed") from None
-        except HTTPError as error:
-            if payload is None and (error.code in (404, 408, 429) or error.code >= 500):
-                return {"status_code": error.code}
-            raise classify_provider_error(error.code, "provider HTTP request failed") from None
-        except (TimeoutError, URLError, OSError) as error:
-            wrapped_timeout = isinstance(error, URLError) and isinstance(error.reason, socket.timeout)
-            kind = ProviderErrorKind.TIMEOUT if isinstance(error, (TimeoutError, socket.timeout)) or wrapped_timeout else ProviderErrorKind.SERVER
-            raise ProviderError(kind, detail=type(error).__name__) from None
-        except (json.JSONDecodeError, UnicodeDecodeError):
-            raise ProviderError(ProviderErrorKind.MALFORMED_RESPONSE, detail="invalid JSON response") from None
-        if not isinstance(parsed, Mapping):
-            raise ProviderError(ProviderErrorKind.MALFORMED_RESPONSE, detail="response must be an object")
-        return parsed
-
-
-@dataclass(frozen=True)
-class ActivityMetadata:
-    activity_id: str
-    submitted_at: datetime
-    endpoint: str
-    request_fields: tuple[str, ...]
-    status_transitions: tuple[str, ...] = ()
-    response_metadata: Mapping[str, object] = field(default_factory=dict)
-
-
-class FortyGuardClient:
-    """Authenticated submit/poll boundary; billable work is submitted exactly once."""
-
-    def __init__(
-        self,
-        transport: FortyGuardTransport,
-        api_key: str,
-        *,
-        clock: Callable[[], datetime],
-        event_sink: Callable[[Mapping[str, object]], None] | None = None,
-        ledger: CreditLedger | None = None,
-    ) -> None:
-        if not api_key:
-            raise ValueError("an API key is required")
-        self._transport = transport
-        self._api_key = api_key
-        self._clock = clock
-        self._event_sink = event_sink
-        self._ledger = ledger
-
-    def submit_and_poll(
-        self,
-        endpoint: str,
-        payload: Mapping[str, object],
-        *,
-        sleep: Callable[[float], None] = default_sleep,
-        max_polls: int = 12,
-    ) -> tuple[Mapping[str, object], ActivityMetadata]:
-        response = self._transport.post(endpoint, payload, self._api_key)
-        status_code = response.get("status_code")
-        if isinstance(status_code, int) and status_code >= 400:
-            raise classify_provider_error(status_code, "activity submission failed")
-        activity_id = response.get("activity_id")
-        if not isinstance(activity_id, str) or not activity_id:
-            raise ProviderError(ProviderErrorKind.MALFORMED_RESPONSE, detail="missing activity id")
-        submitted_at = self._clock()
-        self._emit(
-            "fortyguard.submitted",
-            {"activity_id": activity_id, "endpoint": endpoint, "request": sanitize_payload(payload)},
-        )
-        transitions: list[str] = []
-
-        def get_status(_: str) -> Mapping[str, object]:
-            return self._transport.get(f"/v1/status/{activity_id}", self._api_key)
-
-        result = poll_activity(
-            activity_id,
-            get_status=get_status,
-            sleep=sleep,
-            max_polls=max_polls,
-            on_transition=transitions.append,
-            on_event=self._emit,
-        )
-        metadata = ActivityMetadata(
-            activity_id,
-            submitted_at,
-            endpoint,
-            tuple(sorted(payload)),
-            tuple(transitions),
-            _response_metadata(result),
-        )
-        credits_used = result.get("credits_used")
-        if credits_used is not None:
-            if isinstance(credits_used, bool) or not isinstance(credits_used, int) or credits_used < 0:
-                raise ProviderError(ProviderErrorKind.MALFORMED_RESPONSE, detail="invalid credit usage metadata")
-            if self._ledger is not None:
-                self._ledger.record(UsageRecord(activity_id, endpoint, credits_used, self._clock(), "completed"))
-        self._emit("fortyguard.completed", {"activity_id": activity_id, **_response_metadata(result)})
-        return result, metadata
-
-    def _emit(self, event: str, fields: Mapping[str, object]) -> None:
-        if self._event_sink is not None:
-            self._event_sink({"event": event, "at": self._clock().isoformat(), **fields})
-
-
-def poll_activity(
-    activity_id: str,
-    *,
-    get_status: Callable[[str], Mapping[str, object]],
-    sleep: Callable[[float], None] = default_sleep,
-    max_polls: int = 12,
-    interval_seconds: float = 1.0,
-    on_transition: Callable[[str], None] | None = None,
-    on_event: Callable[[str, Mapping[str, object]], None] | None = None,
-) -> Mapping[str, object]:
-    """Poll one already-submitted billable activity with bounded status checks."""
-    saw_submission_404 = False
-    for poll_number in range(1, max_polls + 1):
-        response = get_status(activity_id)
-        status_code = response.get("status_code")
-        if status_code in (408, 429):
-            if on_transition is not None:
-                on_transition("timed_out" if status_code == 408 else "rate_limited")
-            if on_event is not None:
-                on_event("fortyguard.status_transition", {"activity_id": activity_id, "status": "timed_out" if status_code == 408 else "rate_limited"})
-            if poll_number < max_polls:
-                sleep(interval_seconds)
-                continue
-        if isinstance(status_code, int) and status_code >= 500:
-            if on_transition is not None:
-                on_transition("server_error")
-            if on_event is not None:
-                on_event("fortyguard.status_transition", {"activity_id": activity_id, "status": "server_error"})
-            if poll_number < max_polls:
-                sleep(interval_seconds)
-                continue
-        if status_code == 404 and not saw_submission_404:
-            saw_submission_404 = True
-        elif status_code == 404:
-            raise classify_provider_error(404, "activity not found")
-        status = response.get("status")
-        if isinstance(status, str) and on_transition is not None:
-            on_transition(status)
-        if isinstance(status, str) and on_event is not None:
-            on_event("fortyguard.status_transition", {"activity_id": activity_id, "status": status})
-        if status == "Completed":
-            result = response.get("result")
-            if not isinstance(result, Mapping):
-                raise ProviderError(ProviderErrorKind.MALFORMED_RESPONSE, detail="missing task result")
-            completed = dict(result)
-            for key in ("credits_used", "request_id"):
-                if key in response:
-                    completed[key] = response[key]
-            return completed
-        if status == "Failed":
-            raise ProviderError(ProviderErrorKind.TASK_FAILURE, detail="provider task failed")
-        if status_code not in (None, 200, 202, 404):
-            if not isinstance(status_code, int):
-                raise ProviderError(ProviderErrorKind.MALFORMED_RESPONSE, detail="invalid status code")
-            raise classify_provider_error(status_code, "status lookup failed")
-        if poll_number < max_polls:
-            sleep(interval_seconds)
-    raise ProviderError(ProviderErrorKind.TIMEOUT, detail="activity polling timed out")
-
-
-def _response_metadata(result: Mapping[str, object]) -> dict[str, object]:
-    return {key: result[key] for key in ("credits_used", "request_id") if key in result}
 
 
 @dataclass(frozen=True)
